Allow changing the dev server configuration
I was spending around one hour trying to troubleshot why my devServer config in webpack.config.js didn't work. Turns out, those options are not applied when starting the dev server.

This change implements so it's possible to customize the dev server, if the devServer key is set in the webpack.config.js